backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
transformers
torch
pydantic
